requirements.txt
# python==3.11.9
pyyaml
numpy
scipy
mne
cmcrameri
pyxdf
polars
pandas
PSID
DPAD
